fix(caffe-ffi): 修复 /WX 编译错误 — reshape_layer size_t隐式转换 + test EXPECT_NEAR宏
修复内容：
- src/caffe_ffi/layers/reshape_layer.cpp:70：top_shape.size()(size_t)显式转换为int，消除C4267警告
- tests/cpp/test_objectptr_migration.cpp：28处EXPECT_FLOAT_EQ替换为EXPECT_NEAR(a,b,1e-5f)，使用项目内置测试框架宏

+++ libs/caffe-ffi/src/caffe_ffi/layers/reshape_layer.cpp
@@ -67,7 +67,7 @@ void ReshapeLayer::Reshape(const std::vector<Blob*>& bottom,
     } else if (dim == -1) {
       CAFFE_FFI_CHECK_VALUE_EQ(inferred_axis, -1)
           << "Reshape shape contains multiple -1 dims";
-      inferred_axis = top_shape.size();
+      inferred_axis = static_cast<int>(top_shape.size());
       top_shape.push_back(0);
     } else {
       CAFFE_FFI_CHECK_VALUE_GT(dim, 0) << "Reshape dim must be positive, -1, or 0";
